refactor: move flow logic to flowFactory
- Move flow logic to a different module
- Add xmlHttpRequest w/o jquery implementation
- Use flow factory in both modules
