+++ src/core/fetchRequest.js
@@ -13,69 +13,9 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
- * Copyright (c) 2020-21 (original work) Open Assessment Technologies SA ;
+ * Copyright (c) 2020-2024 (original work) Open Assessment Technologies SA ;
  */
-
-import ApiError from 'core/error/ApiError';
-import NetworkError from 'core/error/NetworkError';
-import TimeoutError from 'core/error/TimeoutError';
-
-const XHR_READY_STATE_OPENED = 1;
-const XHR_READY_STATE_HEADERS_RECEIVED = 2;
-const XHR_READY_STATE_HEADERS_DONE = 4;
-
-/**
- * XHR implementation of Fetch API
- * @param {string} url
- * @param {Object} options - fetch request options that implements RequestInit (https://fetch.spec.whatwg.org/#requestinit)
- * @param {Function} [options.onUploadProgress]
- * @param {Function} [options.onDownloadProgress]
- * @returns {Promise<Response>}
- */
-function xhr(url, options) {
-    return new Promise(resolve => {
-        const request = new XMLHttpRequest();
-        const responseHeaders = new Headers();
-
-        if (typeof options.onUploadProgress === 'function') {
-            request.upload.addEventListener('progress', options.onUploadProgress);
-        }
-        if (typeof options.onDownloadProgress === 'function') {
-            request.addEventListener('progress', options.onDownloadProgress);
-        }
-        request.addEventListener('readystatechange', () => {
-            switch (request.readyState) {
-                case XHR_READY_STATE_OPENED:
-                    for (const header in options.headers) {
-                        request.setRequestHeader(header, options.headers[header]);
-                    }
-                    break;
-                case XHR_READY_STATE_HEADERS_RECEIVED:
-                    request.getAllResponseHeaders().trim().split(/[\r\n]+/).forEach((line) => {
-                        const parts = line.split(": ");
-                        const header = parts.shift();
-                        const value = parts.join(": ");
-                        responseHeaders.append(header, value);
-                    });
-                    break;
-                case XHR_READY_STATE_HEADERS_DONE:
-                    if (request.responseType === 'json') {
-                        request.response = JSON.stringify(xhr.response);
-                    }
-                    const response = new Response(request.response, {
-                        status: request.status,
-                        statusText: request.statusText,
-                        headers: responseHeaders
-                    });
-                    resolve(response);
-                    break;
-            }
-        });
-
-        request.open(options.method, url, true);
-        request.send(options.body);
-    });
-}
+import httpRequestFlowFactory from 'core/request/flowFactory';
 
 /**
  * !!! IE11 requires polyfill https://www.npmjs.com/package/whatwg-fetch
@@ -88,9 +28,6 @@ function xhr(url, options) {
  * @param {integer} [options.timeout] - (default: 5000) if timeout reached, the request will be rejected
  * @param {object} [options.jwtTokenHandler] - core/jwt/jwtTokenHandler instance that should be used during request
  * @param {boolean} [options.returnOriginalResponse] - the full original response should be returned instead of parsing internally (useful for HEAD requests or other empty-response-body requests)
- * @param {'xhr'|'fetch'} [options.httpRequestImplementation] - http request implementation to use
- * @param {Function} [options.onUploadProgress]
- * @param {Function} [options.onDownloadProgress]
  * @returns {Promise<Response>} resolves with http Response object
  */
 const requestFactory = (url, options) => {
@@ -101,103 +38,7 @@ const requestFactory = (url, options) => {
         options
     );
 
-    let flow = Promise.resolve();
-
-    if (options.jwtTokenHandler) {
-        flow = flow
-            .then(options.jwtTokenHandler.getToken)
-            .then(token => ({
-                Authorization: `Bearer ${token}`
-            }))
-            .then(headers => {
-                options.headers = Object.assign({}, options.headers, headers);
-            });
-    }
-
-    const httpRequestImplementation = options.httpRequestImplementation === 'xhr' ? xhr : fetch;
-
-    flow = flow.then(() =>
-        Promise.race([
-            httpRequestImplementation(url, options),
-            new Promise((resolve, reject) => {
-                setTimeout(() => {
-                    reject(new TimeoutError('Timeout', options.timeout));
-                }, options.timeout);
-            })
-        ])
-    );
-
-    if (options.jwtTokenHandler) {
-        flow = flow.then(response => {
-            if (response.status === 401) {
-                return options.jwtTokenHandler
-                    .refreshToken()
-                    .then(options.jwtTokenHandler.getToken)
-                    .then(token => {
-                        options.headers.Authorization = `Bearer ${token}`;
-                        return fetch(url, options);
-                    });
-            }
-
-            return Promise.resolve(response);
-        });
-    }
-
-    /**
-     * Stores the original response
-     */
-    let originalResponse;
-    /**
-     * Stores the response code
-     */
-    let responseCode;
-
-    flow = flow
-        .then(response => {
-            originalResponse = response.clone();
-            responseCode = response.status;
-
-            if (options.returnOriginalResponse) {
-                return originalResponse;
-            }
-            return response.json().catch(() => ({}));
-        })
-        .then(response => {
-            if (responseCode === 204) {
-                return null;
-            }
-
-            // successful request
-            if ((responseCode >= 200 && responseCode < 300) || (response && response.success === true)) {
-                return response;
-            }
-
-            // create error
-            let err;
-            if (response.errorCode) {
-                err = new ApiError(
-                    `${response.errorCode} : ${response.errorMsg || response.errorMessage || response.error}`,
-                    response.errorCode,
-                    originalResponse
-                );
-            } else {
-                err = new NetworkError(
-                    `${responseCode} : Request error`,
-                    responseCode || 0,
-                    originalResponse
-                );
-            }
-            throw err;
-        })
-        .catch(err => {
-            if (!err.type) {
-                //offline, CORS, etc.
-                return Promise.reject(new NetworkError(err.message, 0));
-            }
-            return Promise.reject(err);
-        });
-
-    return flow;
+    return httpRequestFlowFactory(fetch, url, options);
 };
 
 export default requestFactory;
